Run gather step on compute cluster
Since we're running an SGE job to do the prediction,
we may as well have it do the gather step too,
rather than running it in the previous postprocess
(which runs on the server in a single thread).
This also removes the need to have the cluster
scratch disk accessible on the server; compute nodes
likely have faster access to it.